25.11.21 联合巡逻基本完成 补充翻译键 分部界面优化 添加联合巡逻Debug
25.11.20
联巡管理提升至骑士团
常驻骑士职位字典的值改为常驻骑士记录
报错优化、添加[OARO]标识

常驻骑士提升完成
常驻骑士回归改用任务

+++ Branch/BranchMedalHandler.cs
@@ -82,7 +82,7 @@ public class BranchMedalHandler : IExposable
         return 0;
     }
 
-    public void AddMedal(BranchMedalDef medal, short count = 1)
+    public void AddMedal(BranchMedalDef medal, int count = 1)
     {
         if (medal is null)
         {
